# Add Brotli compression provider