--- a/benchmarks/Microsoft.AspNetCore.ResponseCompression.Benchmarks/AssemblyInfo.cs
+++ b/benchmarks/Microsoft.AspNetCore.ResponseCompression.Benchmarks/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+﻿[assembly: BenchmarkDotNet.Attributes.AspNetCoreBenchmark]