--- a/src/Microsoft.AspNetCore.ResponseCompression/BrotliCompressionProvider.cs
+++ b/src/Microsoft.AspNetCore.ResponseCompression/BrotliCompressionProvider.cs
@@ -41,6 +41,7 @@ namespace Microsoft.AspNetCore.ResponseCompression
 #if NETCOREAPP2_1
             return new BrotliStream(outputStream, Options.Level, leaveOpen: true);
 #elif NET461 || NETSTANDARD2_0
+            // Brotli is only supported in .NET Core 2.1+
             throw new PlatformNotSupportedException();
 #else
 #error Target frameworks need to be updated.