--- a/src/Microsoft.AspNetCore.ResponseCompression/ResponseCompressionOptions.cs
+++ b/src/Microsoft.AspNetCore.ResponseCompression/ResponseCompressionOptions.cs
@@ -22,7 +22,8 @@ namespace Microsoft.AspNetCore.ResponseCompression
         public bool EnableForHttps { get; set; } = false;
 
         /// <summary>
-        /// The ICompressionProviders to use for responses.
+        /// The <see cref="ICompressionProvider"/> types to use for responses.
+        /// Providers are prioritized based on the order they are added.
         /// </summary>
         public CompressionProviderCollection Providers { get; } = new CompressionProviderCollection();
     }