--- a/src/Microsoft.AspNetCore.ResponseCompression/ResponseCompressionProvider.cs
+++ b/src/Microsoft.AspNetCore.ResponseCompression/ResponseCompressionProvider.cs
@@ -40,14 +40,14 @@ namespace Microsoft.AspNetCore.ResponseCompression
                 // Use the factory so it can resolve IOptions<GzipCompressionProviderOptions> from DI.
                 _providers = new ICompressionProvider[]
                 {
-                    new CompressionProviderFactory(typeof(GzipCompressionProvider)),
 #if NETCOREAPP2_1
-                    new CompressionProviderFactory(typeof(BrotliCompressionProvider))
+                    new CompressionProviderFactory(typeof(BrotliCompressionProvider)),
 #elif NET461 || NETSTANDARD2_0
                     // Brotli is only supported in .NET Core 2.1+
 #else
 #error Target frameworks need to be updated.
 #endif
+                    new CompressionProviderFactory(typeof(GzipCompressionProvider)),
                 };
             }
             for (var i = 0; i < _providers.Length; i++)