Probably needs a few more tests, like for ordering. Should the Brotli provider be added by default?

Closes https://github.com/aspnet/BasicMiddleware/issues/217