@khellang 2018-06-26
> And at least one that shows what happens when the accept-encoding header includes gzip and br.

Is the intention that gzip will be preferred over Brotli and that you have to explicitly configure it if you want it the other way around?