@khellang 2018-06-27
I guess this will also fix https://github.com/aspnet/BasicMiddleware/issues/216, as I had to take the "registration" order into account when selecting a compression provider when you have multiple matching providers with the same (or missing) quality value.