@shirhatti 2018-07-11
By adding brotli as the default we might have broken anyone running on full Fx. Whoops