@Tratcher 2018-07-11
We do not add brotli on net461.